requirements.txt
langchain-groq
streamlit
sentence-transformers
chromadb
huggingface-hub
python-dotenv
langchain_community